refactoring - move save and load datasets to xr_utils.py.
